+++ R/import_inops.R
@@ -177,6 +177,7 @@ import_inops <- function(
           }
           assign(i, expose[[i]], envir = parent.frame(n = 1))
         }
+        message("Done")
       }
     }