+++ inst/tinytest/regular/test-logic_ops.R
@@ -1,3 +1,5 @@
+
+# basic - equal lengths ====
 x <- c(TRUE, FALSE, TRUE, FALSE)
 y <- c(FALSE, TRUE, TRUE, FALSE)
 outcome <- data.frame(
@@ -12,11 +14,26 @@ x,y,
 "x %?=% y" =c(F,F,F,F),
 check.names = FALSE
 )
-
-
-# "negating logic works (1)"
 expect_equal(outcome, expected)
 
+# basic - unequal lengths ====
+x <- c(TRUE, FALSE)
+y <- c(FALSE, TRUE, TRUE, FALSE)
+outcome <- data.frame(
+  x=x, y=y,
+  "x %xor% y"=x %xor% y, "x %n&% y" = x %n&% y, "x %?=% y" = x %?=% y,
+  check.names = FALSE
+)
+expected <- data.frame(
+  x,y,
+  "x %xor% y"=c(T,T,F,F),
+  "x %n&% y" =c(F,F,F,T),
+  "x %?=% y" =c(F,F,F,F),
+  check.names = FALSE
+)
+expect_equal(outcome, expected)
+
+# negating logic works (1) ====
 df <- expand.grid(x = c(NA, NaN, Inf, -Inf), y = c(NA, NaN, Inf, -Inf))
 x <- df$x
 y <- df$y
@@ -32,68 +49,54 @@ x,y,
 "x %?=% y" = rep(TRUE, 16),
 check.names = FALSE
 )
-# "negating logic works (2)"
 expect_equal(outcome, expected)
 
 
-
+# negating logic works (3) ====
 df <- expand.grid(x = rep(c(NA, NaN, Inf, -Inf), 2), y = rep(c(TRUE, FALSE), 4))
 x <- df$x
 y <- df$y
-# "negating logic works (3)"
 expect_equal(x %?=% y, rep(FALSE, 64))
 
 
 
-# "out works"
+# out works ====
 expect_equal(0:3 %out% 1:10, c(T, F,F,F))
 expect_equal(1:10 %out% 1:3, c(rep(F, 3), rep(T, 7)))
 
 
-
+# numtypes ====
 n <- c(0:5, 0:-5, 0.1, -0.1, 0, 1, Inf, -Inf, NA, NaN)
 cbind(1:length(n), n)
-# "numtype zero"
 expect_equal(c(1e-20, 1) %=numtype% "~0", c(TRUE, FALSE))
 
-# "numtype binary"
 expect_equal(which(n %=numtype% "B"), c(1, 2, 7, 15, 16))
 
-# "numtype prop"
 expect_equal(which(n %=numtype% "prop"), c(1, 2, 7, 13, 15, 16))
 
-# "numtype I"
 expect_equal(which(n %=numtype% "I"), c(1:12, 15:16))
 
-# "numtype I"
 expect_equal(which(n %=numtype% "I"), c(1:12, 15:16))
 
-# "numtype odd"
 expect_equal(which(n %=numtype% "odd"), c(2, 4, 6, 8, 10, 12, 16))
 
-# "numtype even"
 expect_equal(which(n %=numtype% "even"), c(1, 3, 5, 7, 9, 11, 15))
 
-# "numtype real"
 expect_equal(which(n %=numtype% "R"), 1:16)
 
-# "numtype unreal"
 expect_equal(which(n %=numtype% "unreal"), 17:20)
 
 
-
+# strtypes ====
 s <- c(" AbcZ123 ", " abc ", " 1.3 ", " !#$%^&*() ", "", "NA", "NaN", " Inf ")
 cbind(1:length(s), s)
-# "strtype empty"
+
 expect_equal(which(s %=strtype% "empty"), 5)
 
-# "strtype unreal"
 expect_equal(which(s %=strtype% "unreal"), 6:8)
 
-# "strtype numeric"
 expect_equal(which(s %=strtype% "numeric"), c(3, 8))
 
-# "strtype special"
 expect_equal(which(s %=strtype% "special"), 4)
 
 
@@ -122,3 +125,4 @@ expect_error(
   s %=strtype% "foo",
   pattern = "strtype not recognised"
 )
+